[refactor] all the classes earlier in odemis.gui.cont.streams.py have been split and classified.
All classes with base class StreamBarController are in stream_bar.py
StreamController class is in stream.py
