+++ src/odemis/gui/cont/tabs/fastem_chamber_tab.py
@@ -30,7 +30,7 @@ import wx
 from odemis import model
 import odemis.acq.stream as acqstream
 import odemis.gui
-import odemis.gui.cont.streams as streamcont
+from odemis.gui.cont.stream_bar import StreamBarController
 import odemis.gui.cont.views as viewcont
 import odemis.gui.model as guimod
 from odemis.gui.conf.data import get_local_vas
@@ -52,7 +52,7 @@ class FastEMChamberTab(Tab):
             btn.Bind(wx.EVT_TOGGLEBUTTON, self._on_selection_button)
 
         # Create stream & view
-        self._stream_controller = streamcont.StreamBarController(
+        self._stream_controller = StreamBarController(
             tab_data,
             panel.pnl_streams,
             locked=True
