+++ src/odemis/gui/plugin/__init__.py
@@ -33,7 +33,7 @@ import odemis
 from odemis.gui import FG_COLOUR_ERROR, FG_COLOUR_WARNING, FG_COLOUR_MAIN
 from odemis.gui.comp.buttons import ImageTextButton
 from odemis.gui.cont.settings import SettingsController
-from odemis.gui.cont.streams import StreamBarController
+from odemis.gui.cont.stream_bar import StreamBarController
 from odemis.gui.main_xrc import xrcfr_plugin
 from odemis.gui.model import MicroscopeView, MicroscopyGUIData, StreamView
 from odemis.gui.util import call_in_wx_main, get_home_folder
